Adding `Environment` to v1/event
